Move workers outside of pkg/
