+++ worker/thumbnail/thumbnail.go
